refactor: organize code to /src/lib
