--- packages/rsc/examples/basic/src/config-rsc.tsx
@@ -1,21 +1,0 @@
-import { tinyassert } from "@hiogawa/utils";
-import type { BundlerConfig, ImportManifestEntry } from "./react-types";
-
-// bundler id: /src/components/counter.tsx::Counter
-//   ⇓
-// id: /src/components/counter.tsx
-// name: Counter
-
-export const devBundlerConfig: BundlerConfig = new Proxy(
-  {},
-  {
-    get(_target, $$id, _receiver) {
-      console.log("[bundlerConfig]", { $$id });
-      tinyassert(typeof $$id === "string");
-      const [id, name] = $$id.split("::");
-      tinyassert(id);
-      tinyassert(name);
-      return { id, name, chunks: [] } satisfies ImportManifestEntry;
-    },
-  }
-);
