+++ packages/rsc/examples/basic/src/lib/csr.tsx
@@ -1,8 +1,9 @@
 import { once, tinyassert } from "@hiogawa/utils";
-import type { WebpackRequire } from "./react-types";
+import type { WebpackRequire } from "./types";
 
 // __webpack_require__ needs to return stable promise during single render.
-// during dev, vite uses timestamp, so invalidation is not necessary
+// vite uses import with timestamp paramemter during dev,
+// so invalidation is not necessary (hopefully).
 const importOnce = once(clientImport);
 
 const csrWebpackRequire: WebpackRequire = (id) => {
@@ -12,7 +13,6 @@ const csrWebpackRequire: WebpackRequire = (id) => {
 
 async function clientImport(id: string) {
   if (import.meta.env.DEV) {
-    // transformed to `ssrLoadModule` during dev
     return import(/* @vite-ignore */ id);
   } else {
     const clientReferences = await import(
