+++ packages/rsc/examples/basic/src/lib/types-env.d.ts
@@ -1,5 +1,3 @@
-// https://github.com/dai-shi/waku/blob/4d16c28a58204991de2985df0d202f21a48ae1f9/packages/waku/src/types.d.ts#L60-L65
-
 declare module "react-dom/server.edge" {
   export * from "react-dom/server";
 }
@@ -7,7 +5,8 @@ declare module "react-dom/server.edge" {
 declare module "react-server-dom-webpack/server.edge" {
   export function renderToReadableStream(
     node: React.ReactNode,
-    bundlerConfig: import("./react-types").BundlerConfig
+    bundlerConfig: import("./types").BundlerConfig,
+    opitons?: {}
   ): ReadableStream;
 }
 
@@ -15,7 +14,7 @@ declare module "react-server-dom-webpack/client.edge" {
   export function createFromReadableStream(
     stream: ReadableStream,
     options: {
-      ssrManifest: import("./react-types").SsrManifest;
+      ssrManifest: import("./types").SsrManifest;
     }
   ): Promise<React.ReactNode>;
 }
